+++ example/lib/main.dart
@@ -1,5 +1,4 @@
-import 'package:circlify/Circlify.dart';
-import 'package:circlify/circlify_item.dart';
+import 'package:circlify/circlify.dart';
 import 'package:example/animation_demo.dart';
 import 'package:example/custom_value_demo.dart';
 import 'package:flutter/material.dart';